// src/app/interfaces/compartido.ts
export interface Compartido {
    id?: string, //MISMO QUE EN TOKKOBROKER
    app?: number, //VECES QUE EL GUIA COMPARTIO LA PROPIEDAD DESDE LA APP
    facebook?: number, // VECES QUE ALGUIEN COMPARTIO EN FACEBOOK LA PROPIEDAD DESDE LA PÁGINA DEL GUIA
    oportunidades?: number, //VECES QUE EL ALGUIEN INGRESO SUS DATOS EN EL FORMULARIO CREANDO UNA NUEVA OPORTUNIDAD
    visitas?: number, //VECES QUE EL ALGUIEN ENTRO A LA PÁGINA DEL GUÍA
    whatsapp?: number, //VECES QUE EL ALGUIEN COMPARTIO EN WHATSAPP LA PROPIEDAD DESDE LA PÁGINA DEL GUIA
}